requirements.txt
anaconda-anon-usage @ file:///C:/b/abs_1az9wuczyu/croot/anaconda-anon-usage_1749054795139/work
annotated-types @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/annotated-types_1729043709318/work
archspec @ file:///croot/archspec_1709217642129/work
asgiref==3.8.1
boltons @ file:///C:/b/abs_e2_iokhxbp/croot/boltons_1751383740243/work
brotlicffi @ file:///C:/b/abs_f4lixvzmxt/croot/brotlicffi_1736183286104/work
certifi @ file:///C:/b/abs_2bz30lzohz/croot/certifi_1754570654565/work/certifi
cffi @ file:///C:/b/abs_29_b57if3f/croot/cffi_1736184144340/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
colorama @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/colorama_1729036581634/work
colorlog==6.9.0
conda @ file:///C:/b/abs_986_226ly4/croot/conda_1754469528806/work/conda-src
conda-anaconda-telemetry @ file:///C:/b/abs_d4qxn0oeix/croot/conda-anaconda-telemetry_1751062015410/work
conda-anaconda-tos @ file:///C:/b/abs_4csynbis2x/croot/conda-anaconda-tos_1752801584932/work
conda-content-trust @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/conda-content-trust_1729088072778/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1745607008911/work/src
conda-package-handling @ file:///C:/b/abs_7fz3aferfv/croot/conda-package-handling_1731369038903/work
conda_package_streaming @ file:///C:/b/abs_2d41dz5u60/croot/conda-package-streaming_1751307521903/work
cryptography==43.0.3
distro @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/distro_1729059153117/work
Django==4.2
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
frozendict @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/frozendict_1729090952168/work
gunicorn==23.0.0
idna @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/idna_1729038997316/work
inflection==0.5.1
jsonpatch @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/jsonpatch_1729054776004/work
jsonpointer==2.1
libmambapy @ file:///C:/b/abs_627vsv8bhu/croot/mamba-split_1734469608328/work/libmambapy
markdown-it-py @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/markdown-it-py_1729039967492/work
mdurl @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/mdurl_1729039154828/work
menuinst @ file:///C:/b/abs_a0aibbe6s6/croot/menuinst_1750861726761/work
mssql-django==1.5
packaging @ file:///C:/b/abs_3by6s2fa66/croot/packaging_1734472138782/work
platformdirs @ file:///C:/b/abs_ddh15014or/croot/platformdirs_1744273060660/work
pluggy @ file:///C:/b/abs_dfec_m79vo/croot/pluggy_1733170145382/work
pycosat @ file:///C:/b/abs_18nblzzn70/croot/pycosat_1736868434419/work
pycparser==2.22
pydantic @ file:///C:/b/abs_d1rp6fqxgk/croot/pydantic_1750768285689/work
pydantic_core @ file:///C:/b/abs_d0xs0z3erk/croot/pydantic-core_1750754747982/work
Pygments @ file:///C:/b/abs_e4bg5vh5j_/croot/pygments_1744667628203/work
PyJWT==2.10.0
pyodbc==5.2.0
PySocks @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/pysocks_1729039320841/work
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
requests @ file:///C:/b/abs_0b3h9d_724/croot/requests_1750426345668/work
rich @ file:///C:/b/abs_8b92fzmygg/croot/rich_1732638991774/work
ruamel.yaml @ file:///C:/b/abs_0dn1nhdq75/croot/ruamel.yaml_1745960327187/work
ruamel.yaml.clib @ file:///C:/b/abs_475flulhy_/croot/ruamel.yaml.clib_1745937166245/work
setuptools==75.1.0
sqlparse==0.5.1
tqdm @ file:///C:/b/abs_0eh9b6xugj/croot/tqdm_1738945553987/work
truststore @ file:///C:/b/abs_494cm143zh/croot/truststore_1736550137835/work
typing-inspection @ file:///C:/b/abs_261c5jt6kx/croot/typing-inspection_1746023491500/work
typing_extensions @ file:///C:/b/abs_0ffjxtihug/croot/typing_extensions_1734714875646/work
tzdata==2024.2
uritemplate==4.1.1
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
wheel==0.44.0
win_inet_pton @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/win_inet_pton_1729038644686/work
zstandard @ file:///C:/b/abs_31t8xmrv_h/croot/zstandard_1731356578015/work
